logger/requirements.txt
picamera2
sparkfun-qwiic-bme280
python-dotenv
